updating

+++ main.h
@@ -10,7 +10,7 @@ int _printf(const char *format, ...);
 int _putchar(char c);
 int print_str(char *str);
 int print_number(int n);
-int print_binary(unsigned int n);
+int print_binary(unsigned long int n);
 int print_unsi(unsigned long int n, int base);
 int print_hex(unsigned long int n, int uppercase);
 int format_handler(char format, va_list args);
